PlantView Abstraction
Introduce a PlantView class that handles the main information displayed
about a plant. This is then inherited by the UploadResultView so that
the user sees information about their fresh upload. This class will be
inherited by the future UserPlantView and PublicPlantView classes, which handle
the view of an already existing plant based on whether the user is
viewing a plant they own or one that is publicly available and owned by
another user.
