+++ src/main/java/view/upload/UploadResultView.java
@@ -3,6 +3,7 @@ package view.upload;
 import interface_adapter.upload.UploadController;
 import interface_adapter.upload.result.UploadResultState;
 import interface_adapter.upload.result.UploadResultViewModel;
+import view.PlantView;
 import view.ViewComponentFactory;
 
 import javax.swing.*;
@@ -13,137 +14,17 @@ import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
 import java.text.DecimalFormat;
 
-public class UploadResultView extends JPanel implements PropertyChangeListener {
+public class UploadResultView extends PlantView implements PropertyChangeListener {
     private final String viewName = "upload result";
 
     private UploadController controller;
 
-    private final JLabel nameLabel = new JLabel();
-    private final JLabel scientificNameLabel = new JLabel();
-    private final JLabel familyLabel = new JLabel();
-    private final JLabel certaintyLabel = new JLabel();
-    private final JTextArea notesField = new JTextArea();
-    private final JToggleButton togglePublic = new JToggleButton();
-
-    private BufferedImage image;
-
     public UploadResultView(UploadResultViewModel viewModel) {
         viewModel.addPropertyChangeListener(this);
-
-        this.setLayout(new GridBagLayout());
-        this.setBackground(new Color(UploadResultViewModel.TRANSPARENT, true));
-
-        // position each component nicely within the view area using a GridBagLayout
-        GridBagConstraints constraints = new GridBagConstraints();
-        constraints.gridx = 0;
-        constraints.gridy = 0;
-        constraints.gridwidth = 2;
-        constraints.anchor = GridBagConstraints.NORTHWEST;
-        this.add(createTopPanel(), constraints);
-
-        constraints.gridx = 0;
-        constraints.gridy = 1;
-        constraints.gridwidth = 1;
-        constraints.gridheight = 2;
-        constraints.anchor = GridBagConstraints.CENTER;
-        this.add(createImagePanel(), constraints);
-
-        constraints.gridx = 1;
-        constraints.gridy = 1;
-        constraints.gridheight = 1;
-        constraints.fill = GridBagConstraints.BOTH;
-        constraints.anchor = GridBagConstraints.NORTHWEST;
-        constraints.ipady = 20;
-        constraints.weighty = 1;
-        this.add(createContentPanel(), constraints);
-
-        constraints.gridx = 1;
-        constraints.gridy = 2;
-        constraints.ipadx = 20;
-        constraints.weighty = 0;
-        constraints.fill = GridBagConstraints.HORIZONTAL;
-        constraints.anchor = GridBagConstraints.SOUTH;
-        this.add(createActionPanel(), constraints);
     }
 
-    private JPanel createContentPanel() {
-        JPanel contentPanel = new JPanel();
-        SpringLayout layout = new SpringLayout();
-        contentPanel.setLayout(layout);
-        contentPanel.setBackground(new Color(UploadResultViewModel.CONTENT_PANEL_COLOR, true));
-
-        Font font = nameLabel.getFont();
-        nameLabel.setFont(font.deriveFont(Font.BOLD).deriveFont(20f));
-        layout.putConstraint(SpringLayout.WEST, nameLabel, 20, SpringLayout.WEST, contentPanel);
-        layout.putConstraint(SpringLayout.NORTH, nameLabel, 20, SpringLayout.NORTH, contentPanel);
-        contentPanel.add(nameLabel);
-
-        scientificNameLabel.setFont(font.deriveFont(Font.ITALIC).deriveFont(16f));
-        layout.putConstraint(SpringLayout.WEST, scientificNameLabel, 20, SpringLayout.WEST, contentPanel);
-        layout.putConstraint(SpringLayout.NORTH, scientificNameLabel, 5, SpringLayout.SOUTH, nameLabel);
-        contentPanel.add(scientificNameLabel);
-
-        familyLabel.setFont(font.deriveFont(Font.PLAIN).deriveFont(14f));
-        layout.putConstraint(SpringLayout.WEST, familyLabel, 20, SpringLayout.WEST, contentPanel);
-        layout.putConstraint(SpringLayout.NORTH, familyLabel, 5, SpringLayout.SOUTH, scientificNameLabel);
-        contentPanel.add(familyLabel);
-
-        certaintyLabel.setFont(font.deriveFont(Font.PLAIN).deriveFont(14f));
-        layout.putConstraint(SpringLayout.WEST, certaintyLabel, 20, SpringLayout.WEST, contentPanel);
-        layout.putConstraint(SpringLayout.NORTH, certaintyLabel, 15, SpringLayout.SOUTH, familyLabel);
-        contentPanel.add(certaintyLabel);
-
-        notesField.setRows(10);
-        notesField.setFont(font.deriveFont(Font.PLAIN).deriveFont(12f));
-        notesField.setLineWrap(true);
-        notesField.setWrapStyleWord(true);
-        notesField.setMargin(new Insets(10, 10, 10, 10));
-        notesField.setBackground(new Color(UploadResultViewModel.CONTENT_PANEL_COLOR, true));
-        JScrollPane notesScrollPane = new JScrollPane(notesField);
-
-        layout.putConstraint(SpringLayout.WEST, notesScrollPane, 20, SpringLayout.WEST, contentPanel);
-        layout.putConstraint(SpringLayout.EAST, notesScrollPane, -20, SpringLayout.EAST, contentPanel);
-        layout.putConstraint(SpringLayout.NORTH, notesScrollPane, 35, SpringLayout.SOUTH, certaintyLabel);
-        contentPanel.add(notesScrollPane);
-
-        // default privacy to public
-        togglePublic.setFont(font.deriveFont(16f));
-        togglePublic.setMargin(new Insets(0, 0, 0, 0));
-        togglePublic.setPreferredSize(new Dimension(30, 30));
-        togglePublic.setBorder(null);
-        togglePublic.setFocusPainted(false);
-        togglePublic.setText("\uD83D\uDD12");
-        // enable a different color when privacy is toggled
-        togglePublic.setBackground(new Color(UploadResultViewModel.PRIVATE_TOGGLE_COLOR, true));
-        UIManager.put("ToggleButton.select",
-                new Color(UploadResultViewModel.PUBLIC_TOGGLE_COLOR, true));
-        SwingUtilities.updateComponentTreeUI(togglePublic);
-        // enable user feedback when they toggle privacy
-        togglePublic.addItemListener((e) -> {
-            if (e.getStateChange() == ItemEvent.SELECTED) {
-                togglePublic.setText("\uD83D\uDD13");
-            }
-            else {
-                togglePublic.setText("\uD83D\uDD12");
-            }
-        });
-
-        layout.putConstraint(SpringLayout.EAST, togglePublic, -20, SpringLayout.EAST, contentPanel);
-        layout.putConstraint(SpringLayout.NORTH, togglePublic, 20, SpringLayout.NORTH, contentPanel);
-        contentPanel.add(togglePublic);
-
-        return contentPanel;
-    }
-
-    private JPanel createTopPanel() {
-        JPanel topPanel = new JPanel();
-        topPanel.setLayout(new BorderLayout());
-        topPanel.setBackground(new Color(UploadResultViewModel.TOP_PANEL_COLOR, true));
-        topPanel.setPreferredSize(new Dimension(
-                UploadResultViewModel.PANEL_WIDTH,
-                UploadResultViewModel.TOP_PANEL_HEIGHT
-        ));
-
+    protected JPanel createTopPanel() {
+        JPanel topPanel = super.createTopPanel();
         JButton returnBtn = ViewComponentFactory.buildButton(UploadResultViewModel.RETURN_BUTTON_LABEL);
         returnBtn.setBorderPainted(false);
 
@@ -152,44 +33,20 @@ public class UploadResultView extends JPanel implements PropertyChangeListener {
         return topPanel;
     }
 
-    private JPanel createImagePanel() {
-        JPanel imagePanel = new JPanel() {
-            @Override
-            protected void paintComponent(Graphics g) {
-                super.paintComponent(g);
-                if (image != null) {
-                    g.drawImage(image, 0, 0,
-                            UploadResultViewModel.IMAGE_WIDTH,
-                            UploadResultViewModel.IMAGE_HEIGHT,
-                            this
-                    );
-                }
-            }
-        };
-        imagePanel.setPreferredSize(new Dimension(
-                UploadResultViewModel.IMAGE_WIDTH,
-                UploadResultViewModel.IMAGE_HEIGHT
-        ));
-        return imagePanel;
-    }
-
-    private JPanel createActionPanel() {
-        JPanel actionPanel = new JPanel();
-        actionPanel.setLayout(new GridBagLayout());
-        actionPanel.setBackground(new Color(UploadResultViewModel.ACTION_PANEL_COLOR));
+    protected JPanel createActionPanel() {
+        JPanel actionPanel = super.createActionPanel();
 
         JButton saveBtn = ViewComponentFactory.buildButton(UploadResultViewModel.SAVE_BUTTON_LABEL);
         saveBtn.setPreferredSize(new Dimension(100, 30));
 
         saveBtn.addActionListener((e) -> controller.saveUpload(
-                image,
-                scientificNameLabel.getText(),
-                notesField.getText(),
-                togglePublic.isSelected()
+                getImage(),
+                getScientificNameLabel().getText(),
+                getNotesField().getText(),
+                getTogglePublic().isSelected()
         ));
         // NOTE: we negate privacyToggle as the controller needs to know if the image is public,
         //  not if the image is private.
-
         JButton discardBtn = ViewComponentFactory.buildButton(UploadResultViewModel.DISCARD_BUTTON_LABEL);
         discardBtn.setPreferredSize(new Dimension(100, 30));
 
@@ -203,6 +60,7 @@ public class UploadResultView extends JPanel implements PropertyChangeListener {
 
         constraints.gridx = 1;
         actionPanel.add(saveBtn, constraints);
+
         return actionPanel;
     }
 
@@ -215,18 +73,18 @@ public class UploadResultView extends JPanel implements PropertyChangeListener {
     }
 
     private void setFields(UploadResultState state) {
-        this.image = ViewComponentFactory.buildCroppedImage(state.getImagePath());
-        this.nameLabel.setText(state.getName());
-        this.scientificNameLabel.setText(state.getScientificName());
-        this.familyLabel.setText(state.getFamily());
+        this.setImage(ViewComponentFactory.buildCroppedImage(state.getImagePath()));
+        this.getNameLabel().setText(state.getName());
+        this.getScientificNameLabel().setText(state.getScientificName());
+        this.getFamilyLabel().setText(state.getFamily());
 
         DecimalFormat roundingFormat = new DecimalFormat("###.##");
-        this.certaintyLabel.setText(Double.valueOf(roundingFormat.format(state.getCertainty() * 100)) +
+        this.getCertaintyLabel().setText(Double.valueOf(roundingFormat.format(state.getCertainty() * 100)) +
                 "% certainty");
 
-        this.notesField.setText("My notes...");
+        this.getNotesField().setText("My notes...");
 
-        this.togglePublic.setSelected(false);
+        this.getTogglePublic().setSelected(false);
 
         this.revalidate();
         this.repaint();
